requirements.txt
numpy==1.21.0
cloudpickle==1.6.0
importlib_metadata==4.8.1
gym_notices==0.0.4
opencv-python==4.5.3.56
lz4==3.1.3
matplotlib==3.4.3
box2d-py==2.3.5
pygame==2.1.0
ale-py==0.7.1
imageio==2.14.1